first 3 exercises from ssection 1
Exercise 1-3 feels dirty - but - I didn’t know how to return a list that could be used in a function. AND I did not know how to return a defaut in a cond. AND I avoid nested ifs like the plague - which means I am pretty much in trouble for the rest of the book :)